+++ packages/core/bootstrap/src/lib/cache/index.ts
@@ -4,6 +4,7 @@ import { logger } from '../external-adapter'
 import { parseBool, uuid, delay, exponentialBackOffMs, getWithCoalescing } from '../util'
 import * as local from './local'
 import * as redis from './redis'
+import * as metrics from './metrics'
 
 const DEFAULT_CACHE_TYPE = 'local'
 const DEFAULT_CACHE_KEY_GROUP = uuid()
@@ -14,7 +15,7 @@ const DEFAULT_RC_INTERVAL_MAX = 1000
 const DEFAULT_RC_INTERVAL_COEFFICIENT = 2
 const DEFAULT_RC_ENTROPY_MAX = 0
 
-const MAXIMUM_MAX_AGE = 1000 * 60 * 2
+export const MAXIMUM_MAX_AGE = 1000 * 60 * 2
 const ERROR_MAX_AGE = 1000 * 60
 
 const env = process.env
@@ -126,6 +127,7 @@ export const withCache: Middleware = async (execute, options = defaultOptions())
   const _executeWithCache = async (data: AdapterRequest) => {
     const key = _getKey(data)
     const coalescingKey = _getCoalescingKey(key)
+    const endMetrics = metrics.observeMetrics(data.id, key)
     const maxAge = _getRequestMaxAge(data) || _getDefaultMaxAge(data)
     // Add successful result to cache
     const _cacheOnSuccess = async ({ statusCode, data, result }: AdapterResponse) => {
@@ -175,7 +177,22 @@ export const withCache: Middleware = async (execute, options = defaultOptions())
         logger.debug(`Cache: GET ${key}`, entry)
         const reqMaxAge = _getRequestMaxAge(data)
         if (reqMaxAge && reqMaxAge !== entry.maxAge) await _cacheOnSuccess(entry)
-        return { jobRunID: data.id, ...entry }
+        const ttl = await cache.ttl(key)
+        const staleness = entry.maxAge - ttl
+        const debug = {
+          cacheHit: true,
+          staleness,
+          performance: endMetrics(true, staleness),
+          providerCost: 0,
+        }
+        return {
+          jobRunID: data.id,
+          ...entry,
+          debug: {
+            ...(entry.debug || {}),
+            ...debug,
+          },
+        }
       }
       logger.debug(`Cache: SKIP(maxAge < 0)`)
     }
@@ -185,7 +202,12 @@ export const withCache: Middleware = async (execute, options = defaultOptions())
 
     const result = await execute(data)
     await _cacheOnSuccess(result)
-    return result
+    const debug = {
+      staleness: 0,
+      performance: endMetrics(false, 0),
+      providerCost: result.data.cost || 1,
+    }
+    return { ...result, debug: { ...debug, ...result.debug } }
   }
 
   // Middleware wrapped execute fn which cleans up after